requirements.txt
matplotlib==3.9.2
pandas==2.2.3
PyYAML==6.0.2
scikit_learn==1.5.2
seaborn==0.13.2
